* Add pluginutils::ReplaceNative(). * main.cpp: Add a native function hooking example. * plugin.inc.in: Fix an unused variable warning. * pluginutils.h: Move non-inline functions to pluginutils.cpp. * Fix a build error on Linux. * Update copyright year.
